动态设置split time;播放完一个文件能自动跳到下一个文件

+++ src/gstreamer/ReSplic.cpp
@@ -17,7 +17,7 @@
 	 record_handle = CabinInit(1920,1080,30,rgba,1,flip);
 #endif
  }
- void SaveVideoByGST::SetMTime(int Tnm)
+ void SaveVideoByGST::SetMTime(unsigned long Tnm)
  {
 	 SetTime(Tnm,record_handle);
  };
